wip: reformat assignment export code
