+++ src/components/Groups/GroupPage/FeedItem.tsx
@@ -1,33 +1,19 @@
-import { Dialog, Transition } from '@headlessui/react';
+import { Transition } from '@headlessui/react';
 import { CheckIcon, ClipboardListIcon } from '@heroicons/react/outline';
 import { BookmarkIcon } from '@heroicons/react/solid';
-import {
-  arrayUnion,
-  collection,
-  CollectionReference,
-  doc,
-  getFirestore,
-  onSnapshot,
-  query,
-  serverTimestamp,
-  writeBatch,
-} from 'firebase/firestore';
 import { Link } from 'gatsby';
-import React, { Fragment, useContext, useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import toast from 'react-hot-toast';
-import UserDataContext from '../../../context/UserDataContext/UserDataContext';
 import { useActiveGroup } from '../../../hooks/groups/useActiveGroup';
 import { usePostActions } from '../../../hooks/groups/usePostActions';
-import { useUserGroups } from '../../../hooks/groups/useUserGroups';
-import { useFirebaseApp } from '../../../hooks/useFirebase';
 import { GroupData } from '../../../models/groups/groups';
 import {
   getPostTimestampString,
   getTotalPointsOfPost,
   PostData,
 } from '../../../models/groups/posts';
-import { GroupProblemData } from '../../../models/groups/problem';
 import Tooltip from '../../Tooltip/Tooltip';
+import PostExportModal from './PostExportModal';
 
 const AnnouncementIcon = () => {
   return (
@@ -92,126 +78,10 @@ export default function FeedItem({
   isBeingDragged?: boolean;
 }): JSX.Element {
   const { showAdminView } = useActiveGroup();
-  const firebaseApp = useFirebaseApp();
-  const q = query(
-    collection(
-      getFirestore(firebaseApp),
-      'groups',
-      group.id,
-      'posts',
-      post.id,
-      'problems'
-    ) as CollectionReference<GroupProblemData>
-  );
-
-  useEffect(() => {
-    const unsubscribe = onSnapshot(q, {
-      next: snap => {
-        setProblems(snap.docs.map(doc => ({ id: doc.id, ...doc.data() })));
-      },
-      error: error => {
-        toast.error(error.message);
-      },
-    });
-    return () => unsubscribe();
-  }, []);
-
   const { updatePost, deletePost } = usePostActions(group.id);
-
   const [showDropdown, setShowDropdown] = useState(false);
   const [showExportModal, setShowExportModal] = useState(false);
-
-  const { firebaseUser, isLoaded } = useContext(UserDataContext);
-  const groups = useUserGroups();
-  const [problems, setProblems] = React.useState<GroupProblemData[]>([]);
   const ref = React.useRef<HTMLDivElement>();
-
-  const [groupsUsedMap, setGroupsUsedMap] = useState(new Map());
-
-  function handleGroupExportChange(g: GroupData) {
-    console.log(g.name);
-    if (groupsUsedMap.has(g.id)) {
-      setGroupsUsedMap(
-        new Map(
-          groupsUsedMap.set(g.id, new MapData(!groupsUsedMap.get(g.id).used, g))
-        )
-      );
-      console.log(groupsUsedMap.get(g.id).used);
-    } else {
-      setGroupsUsedMap(new Map(groupsUsedMap.set(g.id, new MapData(true, g))));
-      console.log(groupsUsedMap.get(g.id).used);
-    }
-    console.log(groupsUsedMap.size);
-  }
-
-  async function exportSelectedPosts() {
-    const type = post.type;
-    const defaultPost: Omit<PostData, 'timestamp'> = {
-      name: post.name,
-      isPublished: post.isPublished,
-      isPinned: post.isPinned,
-      body: post.body,
-      isDeleted: post.isDeleted,
-      type,
-      pointsPerProblem: post.pointsPerProblem,
-      problemOrdering: post.problemOrdering,
-      ...(type === 'announcement'
-        ? {}
-        : {
-            dueTimestamp: null,
-          }),
-    };
-
-    console.log(groupsUsedMap.size);
-    groupsUsedMap.forEach((value: MapData, key: string) => {
-      if (value.used) {
-        const firestore = getFirestore(firebaseApp);
-        const batch = writeBatch(firestore);
-        const docRef = doc(
-          collection(getFirestore(firebaseApp), 'groups', key, 'posts')
-        );
-
-        batch.set(docRef, { ...defaultPost, timestamp: serverTimestamp() });
-        batch.update(doc(firestore, 'groups', key), {
-          postOrdering: arrayUnion(docRef.id),
-        });
-        batch.commit();
-        // if(post.type != 'announcement') {
-
-        console.log('Problem load ' + problems.length);
-        const batch2 = writeBatch(firestore);
-        problems.map(async problem => {
-          const docRef2 = doc(
-            collection(
-              getFirestore(firebaseApp),
-              'groups',
-              key,
-              'posts',
-              docRef.id,
-              'problems'
-            )
-          );
-
-          problem.id = docRef2.id;
-          problem.isDeleted = false;
-          problem.postId = docRef.id;
-
-          batch2.set(docRef2, { ...(problem as any) });
-          batch2.update(
-            doc(getFirestore(firebaseApp), 'groups', key, 'posts', docRef.id),
-            {
-              [`pointsPerProblem.${docRef2.id}`]: problem.points,
-              [`problemOrdering`]: arrayUnion(docRef2.id),
-            }
-          );
-          console.log(problem);
-        });
-        batch2.commit();
-      }
-    });
-
-    setShowExportModal(false);
-  }
 
   React.useEffect(() => {
     const handleClick = e => {
@@ -396,143 +266,16 @@ export default function FeedItem({
                   </button>
                 </div>
               </Transition>
-
-              <Transition
-                show={showExportModal}
-                enter="transition ease-out duration-100"
-                enterFrom="transform opacity-0 scale-95"
-                enterTo="transform opacity-100 scale-100"
-                leave="transition ease-in duration-75"
-                leaveFrom="transform opacity-100 scale-100"
-                leaveTo="transform opacity-0 scale-95"
-                className="origin-top-right absolute z-10 right-0 mt-2 w-56 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 focus:outline-none"
-                role="menu"
-                aria-orientation="vertical"
-                aria-labelledby="options-menu-0"
-              >
-                <Dialog
-                  as="div"
-                  className="fixed z-10 inset-0 overflow-y-auto"
-                  onClose={() => setShowExportModal(false)}
-                >
-                  <div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
-                    <Transition.Child
-                      as={Fragment}
-                      enter="ease-out duration-300"
-                      enterFrom="opacity-0"
-                      enterTo="opacity-100"
-                      leave="ease-in duration-200"
-                      leaveFrom="opacity-100"
-                      leaveTo="opacity-0"
-                    >
-                      <Dialog.Overlay className="fixed inset-0 bg-gray-500 bg-opacity-75 dark:bg-gray-900 dark:bg-opacity-75 transition-opacity" />
-                    </Transition.Child>
-
-                    {/* This element is to trick the browser into centering the modal contents. */}
-                    <span
-                      className="hidden sm:inline-block sm:align-middle sm:h-screen"
-                      aria-hidden="true"
-                    >
-                      &#8203;
-                    </span>
-                    <Transition.Child
-                      as={Fragment}
-                      enter="ease-out duration-300"
-                      enterFrom="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-                      enterTo="opacity-100 translate-y-0 sm:scale-100"
-                      leave="ease-in duration-200"
-                      leaveFrom="opacity-100 translate-y-0 sm:scale-100"
-                      leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-                    >
-                      <div className="inline-block align-bottom bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-lg sm:w-full">
-                        <div className="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
-                          <div className="sm:flex sm:items-start">
-                            {/*<div className="mx-auto flex-shrink-0 flex items-center justify-center h-12 w-12 rounded-full bg-green-400 sm:mx-0 sm:h-10 sm:w-10">*/}
-                            {/*  <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">*/}
-                            {/*    <path d="M10.894 2.553a1 1 0 00-1.788 0l-7 14a1 1 0 001.169 1.409l5-1.429A1 1 0 009 15.571V11a1 1 0 112 0v4.571a1 1 0 00.725.962l5 1.428a1 1 0 001.17-1.408l-7-14z" />*/}
-                            {/*  </svg>*/}
-                            {/*</div>*/}
-                            <div className="mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left">
-                              <Dialog.Title
-                                as="h3"
-                                className="text-lg leading-6 font-medium text-gray-900"
-                              >
-                                Export Post
-                              </Dialog.Title>
-                              <div className="mt-2 text-gray-500">
-                                Please select the groups you would like to
-                                export this assignment to.
-                                <div className="block">
-                                  {groups.isSuccess &&
-                                    (groups.data?.length > 0 ? (
-                                      groups.data.map(group =>
-                                        // group.ownerIds.includes(firebaseUser.uid)
-                                        group.ownerIds.includes(
-                                          firebaseUser.uid
-                                        ) ? (
-                                          <div key={group.id}>
-                                            <label className="inline-flex items-center">
-                                              <input
-                                                type="checkbox"
-                                                className="form-checkbox"
-                                                onChange={() =>
-                                                  handleGroupExportChange(group)
-                                                }
-                                              />
-                                              <span className="ml-2 text-gray-500">
-                                                {group.name}
-                                              </span>
-                                            </label>
-                                          </div>
-                                        ) : null
-                                      )
-                                    ) : (
-                                      <div>
-                                        <p>
-                                          You aren't in an admin in any groups
-                                          yet!
-                                        </p>
-                                      </div>
-                                    ))}
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                        <div className="bg-gray-50 px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse">
-                          <button
-                            type="button"
-                            className="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-green-600 text-base font-medium text-white hover:bg-gray-400 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 sm:ml-3 sm:w-auto sm:text-sm"
-                            onClick={() => exportSelectedPosts()}
-                          >
-                            Export
-                          </button>
-                          <button
-                            type="button"
-                            className="mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
-                            onClick={() => setShowExportModal(false)}
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      </div>
-                    </Transition.Child>
-                  </div>
-                </Dialog>
-              </Transition>
+              <PostExportModal
+                showExportModal={showExportModal}
+                onClose={() => setShowExportModal(false)}
+                post={post}
+                group={group}
+              />
             </div>
           </div>
         )}
       </div>
     </div>
   );
-}
-
-export class MapData {
-  constructor(b: boolean, g: GroupData) {
-    this.used = b;
-    this.group = g;
-  }
-  used: boolean;
-  group: GroupData;
 }
